Crate lib_word_count

This is a library to count the words in a text.

Words will be sorted by the amount of times they appear in a piece of text.

Modules

pivot_sort
This module contains the implementation of a pivot-based sorting algorithm.

word_index
This module is used to index words and does the heavy lifting of our program.

Functions

count_words
Count the appearances of each word in a string.