Module lib_word_count::pivot_sort

This module contains the implementation of a pivot-based sorting algorithm.

This is used to sort the output of the indexing algorithm, making the final output of the program more readable to humans.

Traits

Sortable
Trait to mark a sortable implementation.

Functions

pivot_sort_high_to_low
Sort a given vector from highest 'weight' to lowest 'weight'.